Make the library allocation-free; drop vendored Urbit headers
SoftBLAS routines operate in place on caller-provided buffers; the only
heap allocation in the compiled library was qasum's accumulator (which
also leaked). Replace it with a stack float128_t. The library now does
zero heap allocation.

The svec/dvec/hvec/qvec helpers (the only other malloc users) are
test-only conveniences — never called by the library or by Vere/Lagoon.
Move them, with float128_pair_t, out of the public header into the test
header, and add NULL checks.

With nothing in the library allocating, the Urbit-allocator coupling is
unnecessary: remove the `#ifdef VERE` block (`#include "noun.h"`,
`#define malloc u3a_malloc`) and delete the 20 vendored Urbit runtime
headers that PR #1 copied into include/. `VERE` was defined nowhere (not
in SoftBLAS's build, not in Vere's ext/softblas build), so the block was
inert; and when SoftBLAS builds inside Vere, Vere's own headers are
already on the include path, so the copies were only a staleness
liability. include/ is now just softblas.h. If the host allocator is
ever needed it can be supplied via a one-line host shim, not a vendored
header tree.

143/143 tests pass; CI builds the same set.
