--- include/noun.h
@@ -1,26 +1,0 @@
-/// @file
-
-#ifndef U3_NOUN_H
-#define U3_NOUN_H
-
-#include "allocate.h"
-#include "hashtable.h"
-#include "error.h"
-#include "jets.h"
-#include "jets/k.h"
-#include "jets/q.h"
-#include "manage.h"
-#include "options.h"
-#include "serial.h"
-#include "types.h"
-#include "vortex.h"
-#include "zave.h"
-#include "imprison.h"
-#include "log.h"
-#include "nock.h"
-#include "retrieve.h"
-#include "trace.h"
-#include "urth.h"
-#include "xtract.h"
-
-#endif /* ifndef U3_NOUN_H */
