+++ include/softblas.h
@@ -4,15 +4,6 @@
 #include "softfloat.h"
 #include <stdlib.h>
 
-/*  If we are building for Urbit, we need to use the Urbit allocator.
-    You are responsible for linking SoftBLAS in that case.
-*/
-#ifdef VERE
-#include "noun.h"
-#define malloc u3a_malloc
-#define free u3a_free
-#endif
-
 //  TYPES
 
 typedef struct {
@@ -370,95 +361,6 @@ static inline void nan_unify_v(complex128_t* a) {
     }
 }
 
-/*  Convert an array of half-precision floats to an array of float16_t.
-    The array of floats is assumed to be in little-endian format.
-
-    Example usage:
-        float16_t* HX = hvec((uint16_t[]){
-            0x3C00,  // 1.0
-            0xBC00,  // -1.0
-            0x4248,  // 3.14
-        }, 3);
-        free(HX);
- */
-static inline float16_t* hvec(uint16_t values[], uint64_t size) {
-    float16_t* result = malloc(size * sizeof(float16_t));
-    
-    for (uint64_t i = 0; i < size; i++) {
-        result[i].v = values[i];
-    }
-    
-    return result;
-}
-
-/*  Convert an array of single-precision floats to an array of float32_t.
-    The array of floats is assumed to be in little-endian format. 
-
-    Example usage:
-        float32_t* SX = svec((float[]){
-            1.0f,
-            -1.0f,
-            3.14159274f
-        }, 3);
-        free(SX);
- */
-static inline float32_t* svec(float values[], uint64_t size) {
-    float32_t* result = malloc(size * sizeof(float32_t));
-    
-    for (uint64_t i = 0; i < size; i++) {
-        result[i].v = *(uint32_t*)&values[i];
-    }
-    
-    return result;
-}
-
-/*  Convert an array of double-precision floats to an array of float64_t.
-    The array of floats is assumed to be in little-endian format. 
-
-    Example usage:
-        float64_t* DX = dvec((double[]){
-            1.0,
-            -1.0,
-            3.1415926535897931
-        }, 3);
-        free(DX);
- */
-static inline float64_t* dvec(double values[], uint64_t size) {
-    float64_t* result = malloc(size * sizeof(float64_t));
-    
-    for (uint64_t i = 0; i < size; i++) {
-        result[i].v = *(uint64_t*)&values[i];
-    }
-    
-    return result;
-}
-
-typedef struct {
-    uint64_t hi;
-    uint64_t lo;
-} float128_pair_t;
-
-/*  Convert an array of quad-precision float pairs to an array of float128_t.
-    The array of floats is assumed to be in little-endian format. 
-
-    Example usage:
-        float128_t* QX = qvec((float128_pair_t[]){
-            {.hi = 0x3FF0000000000000, .lo = 0x0000000000000000},  // 1.0
-            {.hi = 0xBFF0000000000000, .lo = 0x0000000000000000},  // -1.0
-            {.hi = 0x400921FB54442D18, .lo = 0x469898cc51701b80},  // 3.141592653589793238462643383279502797479068098137295573004504331874296718662975536062731407582759857177734375
-        }, 3);
-        free(QX);
- */
-static inline float128_t* qvec(float128_pair_t pairs[], uint64_t size) {
-    float128_t* result = malloc(size * sizeof(float128_t));
-    
-    for (uint64_t i = 0; i < size; i++) {
-        result[i].v[0] = pairs[i].lo;
-        result[i].v[1] = pairs[i].hi;
-    }
-    
-    return result;
-}
 
   static inline void
   _set_rounding(uint_fast8_t a)
